NYC COVID DEATHS
For this project, I decided to use the NYC OpenData. I chose “COVID-19  highest deaths”. Here are the necessary links. 

https://data.cityofnewyork.us/Health/COVID-19-Daily-Counts-of-Cases-Hospitalizations-an/rc75-m7u3

https://data.cityofnewyork.us/resource/rc75-m7u3.json

I chose this API because they are free and public. First, I fetched all the JSON data, then focused only on date and death count. After that, I sorted the data to show me the highest death count and plotted a bar chart to show the result.